refactor: move shared kernel utility functions to utils package
