+++ stocking/src/main/scala/dev/atedeg/mdm/stocking/Actions.scala
@@ -5,7 +5,8 @@ import cats.data.{ NonEmptyList, NonEmptySet }
 import cats.syntax.all.*
 
 import dev.atedeg.mdm.products.*
-import dev.atedeg.mdm.products.given
+import dev.atedeg.mdm.products.utils.*
+import dev.atedeg.mdm.products.utils.given
 import dev.atedeg.mdm.stocking.Errors.*
 import dev.atedeg.mdm.stocking.OutgoingEvent.*
 import dev.atedeg.mdm.utils.*
