+++ stocking/src/test/scala/dev/atedeg/mdm/stocking/Tests.scala
@@ -9,6 +9,7 @@ import org.scalatest.featurespec.AnyFeatureSpec
 import org.scalatest.matchers.should.Matchers
 
 import dev.atedeg.mdm.products.*
+import dev.atedeg.mdm.products.utils.*
 import dev.atedeg.mdm.stocking.Errors.*
 import dev.atedeg.mdm.stocking.OutgoingEvent.*
 import dev.atedeg.mdm.utils.*
